requirements.txt
-f /usr/share/pip-wheels
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-crispy-forms==1.11.2
django-extensions==3.1.2
django-filter==2.4.0
django-guardian==2.4.0
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
idna==2.10
importlib-metadata==3.10.0
Markdown==3.3.4
mysqlclient==2.0.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
